Remove email from .gitconfig
I had thought that I could just use an environment variable in the
.gitconfig file, but I was wrong. However, Git responds to
GIT_AUTHOR_EMAIL as an environment varaible. I just set this in a bash
profile file, and that worked. Note: I cannot seem to override this
environment variable, even if I set an email address in a global or
project-specific Git configuration file.